src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/attribute/primitive/SinglePrimitiveBooleanAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.attribute.primitive;

import com.github.brunomndantas.jscrapper.support.parser.single.attribute.SingleAttributeParser;
import org.openqa.selenium.WebElement;

public class SinglePrimitiveBooleanAttributeParser extends SingleAttributeParser {

    public SinglePrimitiveBooleanAttributeParser(String attribute) {
        super(attribute, false);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Boolean.parseBoolean(element.getAttribute(super.attribute));
    }

}
